Introduce fragile marker for text document
This patch introduces fragile and sticky marker for text document. Fragile
markers are identical to sticky markers except for fragile markers are
removed when characters in fragile markers are changed, e.g. inserting or
deleting characters in a range of fragile marker.

From this patch, we call existing markers to "Sticky marker."

BUG=n/a
TEST=MarkerSetTest.DidDeleteAtFragile
TEST=MarkerSetTest.DidInsertAtFragile